+++ Tests/SentryTests/Integrations/SentryAppStartTrackerTests.swift
@@ -12,7 +12,7 @@ class SentryAppStartTrackerTests: XCTestCase {
         let currentDate = TestCurrentDateProvider()
         let sysctl = TestSysctl()
         let fileManager: SentryFileManager
-        let crashAdapter = TestSentryCrashWrapper()
+        let crashAdapter = TestSentryCrashAdapter.sharedInstance()
         let appStateManager: SentryAppStateManager
         
         let appStartDuration: TimeInterval = 0.4
@@ -198,21 +198,21 @@ class SentryAppStartTrackerTests: XCTestCase {
     private func sendAppMeasurement() {
         SentrySDK.getAndResetAppStartMeasurement()
     }
-    
+
     private func assertValidStart(type: SentryAppStartType) {
         let appStartMeasurement = SentrySDK.getAndResetAppStartMeasurement()
         XCTAssertNotNil(appStartMeasurement)
         XCTAssertEqual(type.rawValue, appStartMeasurement?.type.rawValue)
-        
+
         let expectedAppStartDuration = fixture.appStartDuration
         let actualAppStartDuration = appStartMeasurement!.duration
         XCTAssertEqual(expectedAppStartDuration, actualAppStartDuration, accuracy: 0.000_1)
-        
+
         XCTAssertEqual(fixture.sysctl.processStartTimestamp, appStartMeasurement?.appStartTimestamp)
         XCTAssertEqual(fixture.runtimeInitTimestamp, appStartMeasurement?.runtimeInitTimestamp)
-        
+
     }
-    
+
     private func assertNoAppStartUp() {
         XCTAssertNil(SentrySDK.getAndResetAppStartMeasurement())
     }