+++ Tests/SentryTests/SentryOptionsTest.m
@@ -367,27 +367,6 @@
     XCTAssertEqual(NO, options.enableOutOfMemoryTracking);
 }
 
-- (void)testEnableAppStartMeasuring
-{
-    SentryOptions *options = [self getValidOptions:@{ @"enableAppStartMeasuring" : @YES }];
-
-    XCTAssertEqual(YES, options.enableAppStartMeasuring);
-}
-
-- (void)testDefaultAppStartMeasuring
-{
-    SentryOptions *options = [self getValidOptions:@{}];
-
-    XCTAssertEqual(YES, options.enableAppStartMeasuring);
-}
-
-- (void)testSetEnableAppStartMeasuringGargabe
-{
-    SentryOptions *options = [self getValidOptions:@{ @"enableAppStartMeasuring" : @"" }];
-
-    XCTAssertEqual(NO, options.enableAppStartMeasuring);
-}
-
 - (void)testSessionTrackingIntervalMillis
 {
     NSNumber *sessionTracking = @2000;
@@ -437,7 +416,6 @@
     XCTAssertEqual(@1, options.sampleRate);
     XCTAssertEqual(YES, options.enableAutoSessionTracking);
     XCTAssertEqual(YES, options.enableOutOfMemoryTracking);
-    XCTAssertEqual(YES, options.enableAppStartMeasuring);
     XCTAssertEqual([@30000 unsignedIntValue], options.sessionTrackingIntervalMillis);
     XCTAssertEqual(YES, options.attachStacktrace);
     XCTAssertEqual(20 * 1024 * 1024, options.maxAttachmentSize);
@@ -610,6 +588,36 @@
     SentryOptions *options = [self getValidOptions:@{}];
 
     XCTAssertNil(options.tracesSampler);
+}
+
+- (void)testIsTracingEnabled_NothingSet_IsDisabled
+{
+    SentryOptions *options = [[SentryOptions alloc] init];
+    XCTAssertFalse(options.isTracingEnabled);
+}
+
+- (void)testIsTracingEnabled_TracesSampleRateSetToZero_IsDisabled
+{
+    SentryOptions *options = [[SentryOptions alloc] init];
+    options.tracesSampleRate = @0.00;
+    XCTAssertFalse(options.isTracingEnabled);
+}
+
+- (void)testIsTracingEnabled_TracesSampleRateSet_IsEnabled
+{
+    SentryOptions *options = [[SentryOptions alloc] init];
+    options.tracesSampleRate = @0.01;
+    XCTAssertTrue(options.isTracingEnabled);
+}
+
+- (void)testIsTracingEnabled_TracesSamplerSet_IsEnabled
+{
+    SentryOptions *options = [[SentryOptions alloc] init];
+    options.tracesSampler = ^(SentrySamplingContext *context) {
+        XCTAssertNotNil(context);
+        return @0.0;
+    };
+    XCTAssertTrue(options.isTracingEnabled);
 }
 
 - (void)testInAppIncludes