+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Unreleased
+
+- feat: Support XCFramework for Carthage (#1175)
+- Remove invalid excludes from `Package.swift` (#1169)
+
+## 7.2.0-beta.1
+
+- feat: Measure slow and frozen frames (#1123)
+- fix: Operation names for auto instrumentation (#1164)
+
 ## 7.2.0-beta.0
 
 - fix: Set description for performance tracker spans (#1042)